refactor: move lexer to parser mod
