+++ crates/yajl/src/parser/lexer.rs
